MAINT: rename new module, scope expanded
